experiments/hctest/requirements.txt
hcaptcha-challenger>=0.18.0
playwright>=1.40.0
python-dotenv>=1.0.0